refactor(#3270): extract method-trampolines subsystem from closures.ts
Verbatim move of the method-ABI→closure-ABI trampoline machinery (~788 LOC): object-literal/cached method closures, pending-trampoline finalize pass, null-this TypeError + this-slot prologue helpers. Byte-identity IDENTICAL 39/39, tsc 0. #3182.

✓
